website-qa-bot/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pinecone-client
requests
